Messing around with getting last match stats for a player Update readme
