+++ user-proflie.html
@@ -1,1 +1,175 @@
-<!-- TODO -->
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>didtheyplay.soccer</title>
+    <link rel="stylesheet" href="css/style.css" />
+  </head>
+  <body>
+    <header>
+      <div class="header-title">
+        <img class="header-image" src="img/ball.png" />
+        <h1>didtheyplay.soccer?</h1>
+
+        <!-- THESE ARE FOR THE LOGIN AND REGISTER BUTTONS -->
+        <div class="login-register-buttons">
+          <button
+            onclick="document.getElementById('login-form').style.display='block'"
+            style="width: auto"
+          >
+            Login
+          </button>
+          <button
+            onclick="document.getElementById('register-form').style.display='block'"
+            style="width: auto"
+          >
+            Register
+          </button>
+        </div>
+      </div>
+      <nav class="header-nav">
+        <ul>
+          <li><a href="index.html" id="this">Home</a></li>
+          <li><a href="players.html">Players</a></li>
+          <li><a href="fixtures.html">Fixtures</a></li>
+          <li><a href="tables.html">Tables</a></li>
+        </ul>
+      </nav>
+    </header>
+    <main>
+      <section>
+        <h2 style="text-align: center">Welcome, User Name</h2>
+        <div class="user-favorite-players">
+          <h3>Favorite Players</h3>
+          <ul>
+            <!--This list should be populated by the users fav players in the database-->
+            <li>Matthew Hoppe</li>
+            <li>Weston McKennie</li>
+            <li>Pedri</li>
+            <li>Eric Garcia</li>
+            <li>Phil Foden</li>
+          </ul>
+        </div>
+        <div class="favorite-player-display">
+          <!--selected player last match stats go here-->
+          <article></article>
+        </div>
+      </section>
+
+      <!-- THESE ARE FOR THE LOGIN AND REGISTER BUTTONS -->
+      <section class="login-and-register">
+        <div id="login-form" class="login-window">
+          <form class="login-window-box animate" action="" method="post">
+            <div class="close-x-div">
+              <span
+                onclick="document.getElementById('login-form').style.display='none'"
+                class="close"
+                >&times;</span
+              >
+            </div>
+
+            <div class="login-info">
+              <label for="username"><b>Username</b></label>
+              <input
+                type="text"
+                placeholder="Enter Username"
+                name="username"
+                required
+              />
+
+              <label for="password"><b>Password</b></label>
+              <input
+                type="password"
+                placeholder="Enter Password"
+                name="password"
+                required
+              />
+
+              <button type="submit">Login</button>
+              <button type="reset" class="clearButton">Clear</button>
+            </div>
+
+            <div class="container" style="background-color: #f1f1f1">
+              <span>Are you not <a href="">registered?</a></span>
+            </div>
+          </form>
+        </div>
+
+        <div id="register-form" class="register-window">
+          <form class="register-window-box animate" action="" method="post">
+            <div class="close-x-div">
+              <span
+                onclick="document.getElementById('register-form').style.display='none'"
+                class="close"
+                >&times;</span
+              >
+            </div>
+
+            <div class="register-info">
+              <label for="username"><b>Username</b></label>
+              <input
+                type="text"
+                placeholder="Enter Username"
+                name="username"
+                required
+              />
+
+              <label for="password"><b>Password</b></label>
+              <input
+                type="password"
+                placeholder="Enter Password"
+                name="password"
+                required
+              />
+
+              <label for="email"><b>E-Mail</b></label>
+              <input
+                type="email"
+                placeholder="email@domain.com"
+                name="email"
+                required
+              />
+
+              <label for="phone"><b>Telephone</b></label>
+              <input
+                type="tel"
+                placeholder="(###) ###-####"
+                name="password"
+                pattern="\(\d{3}\) +\d{3}-\d{4}"
+                required
+              />
+
+              <label for="birthdate"><b>Date of Birth</b></label>
+              <input
+                type="date"
+                placeholder="mm/dd/yyyy"
+                name="birthdate"
+                required
+              />
+
+              <button type="submit">Register</button>
+              <button type="reset" class="clearButton">Clear</button>
+            </div>
+
+            <div class="container" style="background-color: #f1f1f1">
+              <span>Are you already <a href="">registered?</a></span>
+            </div>
+          </form>
+        </div>
+      </section>
+    </main>
+    <footer>
+      <div>
+        All statistics provided by
+        <a href="http://www.api-football.org">api-football</a>.
+      </div>
+      <div>
+        Some icons made by
+        <a href="https://www.freepik.com" title="Freepik">Freepik</a> from
+        <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a
+        >.
+      </div>
+    </footer>
+  </body>
+</html>
